requirements.txt
Flask==3.1.0
openai==1.76.0
protobuf==6.30.2
python-dotenv==1.1.0
gunicorn
google-cloud-dialogflow
google-auth>=2.0.0
flask-cors